// src/api/generated/model/updateCreativeTaskRequestDtoOrdFlagsItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Title
 * Description
 * OpenAPI spec version: 1.0
 */

export type UpdateCreativeTaskRequestDtoOrdFlagsItem = typeof UpdateCreativeTaskRequestDtoOrdFlagsItem[keyof typeof UpdateCreativeTaskRequestDtoOrdFlagsItem];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UpdateCreativeTaskRequestDtoOrdFlagsItem = {
  social: 'social',
  native: 'native',
  social_quota: 'social_quota',
} as const;