// src/api/generated/model/updateCreativeTaskRequestDtoOrdForm.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Title
 * Description
 * OpenAPI spec version: 1.0
 */

/**
 * Форма распространения креатива (тип креатива)
 * @nullable
 */
export type UpdateCreativeTaskRequestDtoOrdForm = typeof UpdateCreativeTaskRequestDtoOrdForm[keyof typeof UpdateCreativeTaskRequestDtoOrdForm] | null;


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UpdateCreativeTaskRequestDtoOrdForm = {
  banner: 'banner',
  text_block: 'text_block',
  text_graphic_block: 'text_graphic_block',
  audio: 'audio',
  video: 'video',
  live_audio: 'live_audio',
  live_video: 'live_video',
  text_video_block: 'text_video_block',
  text_graphic_video_block: 'text_graphic_video_block',
  text_audio_block: 'text_audio_block',
  text_graphic_audio_block: 'text_graphic_audio_block',
  text_audio_video_block: 'text_audio_video_block',
  text_graphic_audio_video_block: 'text_graphic_audio_video_block',
  banner_html5: 'banner_html5',
} as const;