// src/api/generated/model/updateCreativeTaskRequestDtoOrdPayType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Title
 * Description
 * OpenAPI spec version: 1.0
 */

/**
 * Тип оплаты креатива (pay_type в ОРД)
 * @nullable
 */
export type UpdateCreativeTaskRequestDtoOrdPayType = typeof UpdateCreativeTaskRequestDtoOrdPayType[keyof typeof UpdateCreativeTaskRequestDtoOrdPayType] | null;


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UpdateCreativeTaskRequestDtoOrdPayType = {
  cpm: 'cpm',
  cpc: 'cpc',
  cpa: 'cpa',
  other: 'other',
} as const;